+++ bika/lims/idserver.py
@@ -35,6 +35,7 @@ from bika.lims.interfaces import IAnalysisRequestSecondary
 from bika.lims.interfaces import IIdServer
 from bika.lims.numbergenerator import INumberGenerator
 from DateTime import DateTime
+from datetime import datetime
 from Products.ATContentTypes.utils import DT2dt
 from zope.component import getAdapters
 from zope.component import getUtility
@@ -206,6 +207,7 @@ def get_variables(context, **kw):
         "id": api.get_id(context),
         "portal_type": portal_type,
         "year": get_current_year(),
+        "yymmdd": get_yymmdd(),
         "parent": api.get_parent(context),
         "seq": 0,
         "alpha": Alphanumber(0),
@@ -341,6 +343,10 @@ def get_current_year():
     """
     return DateTime().strftime("%Y")[2:]
 
+def get_yymmdd():
+    """Returns the current date in yymmdd format
+    """
+    return datetime.now().strftime("%y%m%d")
 
 def make_storage_key(portal_type, prefix=None):
     """Make a storage (dict-) key for the number generator